+++ src/media/color/mod.rs
@@ -6,6 +6,7 @@
 #![cfg_attr(feature = "safe_color", forbid(unsafe_code))]
 
 mod base; // trait ColorBase
+mod luminance; // Luminance
 mod namespace; // struct Color
 mod rgb; // Rgb[a][8|16|F32|F64], Rgb[a]Lin[F32|F64]
 
@@ -16,9 +17,8 @@ crate::items! { // structural access: _mods, _all, _always
     pub use _always::*;
 
     mod _mods {
-        pub use super::{base::*, namespace::*, rgb::_all::*};
+        pub use super::{base::*, luminance::*, namespace::*, rgb::_all::*};
         // WIPZONE
-        // pub use super::gray::*;
         // pub use super::spectral::*;
     }
     pub(super) mod _all {
@@ -29,5 +29,4 @@ crate::items! { // structural access: _mods, _all, _always
     }
 }
 // WIPZONE
-// mod gray;
 // mod spectral;